+++ core/domain/training_job_result_domain_test.py
@@ -16,7 +16,9 @@
 
 """Tests for gce metadata services"""
 
+from core.classifiers import algorithm_registry
 from core.domain import training_job_result_domain
+from core.domain.proto import training_job_response_payload_pb2
 from core.domain.proto import text_classifier_pb2
 from core.tests import test_utils
 
@@ -48,3 +50,41 @@ class TrainingJobResultTests(test_utils.GenericTestBase):
             Exception,
             'Expected classifier data of type TextClassifier'):
             job_result.validate()
+
+    def test_that_all_algorithms_have_job_result_information(self):
+        """Test that all algorithms have properties to identify name and type
+        of attribute in job result proto which stores classifier data for that
+        algorithm.
+        """
+        job_result_proto = (
+            training_job_response_payload_pb2.
+            TrainingJobResponsePayload.JobResult())
+        for classifier in algorithm_registry.Registry.get_all_classifiers():
+            self.assertIsNotNone(classifier.name_in_job_result_proto)
+            attribute_type_name = type(getattr(
+                job_result_proto, classifier.name_in_job_result_proto)).__name__
+            self.assertEqual(
+                attribute_type_name, classifier.type_in_job_result_proto)
+
+    def test_that_training_job_result_proto_is_generated_with_correct_details(
+            self):
+        """Ensure that the JobResult proto is correctly generated from
+        TrainingJobResult domain object.
+        """
+        classifier_data = text_classifier_pb2.TextClassifierFrozenModel()
+        classifier_data.model_json = 'dummy model'
+        job_id = 'job_id'
+        algorithm_id = 'TextClassifier'
+        classifier = algorithm_registry.Registry.get_classifier_by_algorithm_id(
+            algorithm_id)
+        job_result = training_job_result_domain.TrainingJobResult(
+            job_id, algorithm_id, classifier_data)
+        job_result_proto = job_result.to_proto()
+
+        # Lint test for no-member needs to be disabled as protobuf generated
+        # classes are metaclasses and hence their attributes are defined at
+        # runtime.
+        self.assertEqual(job_result_proto.job_id, job_id) # pylint: disable=no-member
+        self.assertEqual(
+            job_result_proto.WhichOneof('classifier_frozen_model'),
+            classifier.name_in_job_result_proto)
